adjust CronjobStore#getInitialDelayInMilliSeconds after r3409: changed initialDelay and stopTimeout to long to avoid losing precision
git-svn-id: https://svn.exedio.com/tools/cronjob/trunk@3412 8ceed453-710c-0410-87a1-99b3265506e7

+++ examplesrc/com/exedio/cronjob/example/ExampleStore.java
@@ -50,7 +50,7 @@ public final class ExampleStore implements CronjobStore
 				new NullNameJob());
 	}
 
-	public int getInitialDelayInMilliSeconds()
+	public long getInitialDelayInMilliSeconds()
 	{
 		System.out.println("ExampleStore.getInitialDelayInMilliSeconds");
 		return 1000;
